auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Servicio de autenticación de usuarios con FastAPI y MongoDB"
authors = [
    { name = "Matias Vagliviello", email = "vaglimatias@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.120.2,<0.121.0",
    "uvicorn>=0.38.0,<0.39.0",
    "motor>=3.6.0,<4.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "email-validator>=2.0.0,<3.0.0",
    "python-multipart>=0.0.6,<0.1.0",
    "bcrypt==3.2.2",
    "pymongo>=4.15.3,<5.0.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
        "pika>=1.3.2,<2.0.0",
    "python-jose>=3.5.0,<4.0.0",
    "celery>=5.3.6,<6.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false


[dependency-groups]
dev = [
    "pytest (<9)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
